move fatsecret_authenticate to fatsecret.py
